yamiencoder: use codec name or extension for output creation
codec name is an accepted command line parameter and it can be use
to decide on the codec to create.  If it is not given then the
codec is still decided based on the extension.

VP8 and VP9 generate ivf content, so codec name is necessary to
make a decision

Signed-off-by: Daniel Charles <daniel.charles@intel.com>