yamitranscode: make vp9 encoder work with yamitranscode
yamitranscode is not honoring the codec name on the command line
so it is not possible to know which codec is intended for the output
if a common extension is given. So far, common extension is ivf and it
is required that yamitranscode calls EncodeOutput->create with a codec name
for examples to run properly

supported examples below, the output file is modified on this patch. The
rest of the command line options can be modified at will

yamitranscode -i ./flambert_main.mp4_1920x1080.I420 -W 1920 -H 1080 --intraperiod 300 \
--rcmode CBR -b 2000 -o ./flambert_kfdist300_2Mbps_ytranscode.vp9

yamitranscode -i ./flambert_main.mp4_1920x1080.I420 -W 1920 -H 1080 --intraperiod 300 \
--rcmode CBR -b 2000 -o -c VP9 ./flambert_kfdist300_2Mbps_ytranscode.ivf

Signed-off-by: Daniel Charles <daniel.charles@intel.com>